Adding back the original rake completion
Need to smack Steve Losh for deleting it along with a bunch of other
aliases, made me loose my history. I did spend some time trying to pick
only the changes I wanted -- but this was removed in the same commit and
after an hour I gave up and did it the brute force way.

This would have been easier if I didn't like all the other modifications
Steve made.